[msx2] new separate vblank function
